// packages/core/src/types.ts
export enum SupervisionDirective {
  Resume = 'resume',
  Restart = 'restart',
  Stop = 'stop',
  Escalate = 'escalate',
}
// ...
export enum RoutingStrategy {
  RoundRobin = 'round-robin',
  Random = 'random',
  SmallestMailbox = 'smallest-mailbox',
  Broadcast = 'broadcast',
  ConsistentHash = 'consistent-hash',
}